+++ src/back_end/models/userModel.js
@@ -1,8 +1,6 @@
-// load the things we need
 var mongoose = require('mongoose');
 var bcrypt   = require('bcrypt-nodejs');
 
-// define the schema for our user model
 var userSchema = mongoose.Schema({
     username: String,
     password: String,
@@ -16,16 +14,12 @@ var userSchema = mongoose.Schema({
     }
 });
 
-// methods ======================
-// generating a hash
 userSchema.methods.generateHash = function(password) {
     return bcrypt.hashSync(password, bcrypt.genSaltSync(8), null);
 };
 
-// checking if password is valid
 userSchema.methods.validPassword = function(password) {
     return bcrypt.compareSync(password, this.password);
 };
 
-// create the model for users and expose it to our app
 module.exports = mongoose.model('User', userSchema);
