+++ src/front-end/js/controllers/authController.js
@@ -1,12 +1,9 @@
 Harley.controller("AuthModalController", ["$scope", "$uibModal", "LoginFactory", function ($scope, $uibModal, LoginFactory) {
     $scope.authModalOpen = function () {
-        console.log("Modal will be open here!!!");
-
         $scope.modalInstance = $uibModal.open({
             size: "sm",
             templateUrl: "views/authModal.html",
             controller: ["$scope", "$uibModalInstance", function ($scope, $uibModalInstance) {
-                console.log($uibModalInstance);
                 $scope.ok = function () {
                     console.log($scope.modalText);
                     $uibModalInstance.close({modalText: $scope.modalText});
@@ -19,15 +16,12 @@ Harley.controller("AuthModalController", ["$scope", "$uibModal", "LoginFactory",
         });
 
         $scope.modalInstance.result.then(function (data) {
-            console.log("Ok");
-            console.log(data);
             $scope.modalText = data["modalText"];
         }, function () {
             console.log("Reject");
         });
 
         $scope.login = function (user) {
-            console.log("LoginCtrl", user);
             LoginFactory.post(user);
         };
     };
@@ -35,7 +29,6 @@ Harley.controller("AuthModalController", ["$scope", "$uibModal", "LoginFactory",
         { title:'Dynamic Title 1', content:'Dynamic content 1' },
         { title:'Dynamic Title 2', content:'Dynamic content 2', disabled: true }
     ];
-
     $scope.model = {
         name: 'Tabs'
     };
@@ -58,10 +51,8 @@ Harley.controller("SignUpCtrl", ["$scope", "$rootScope", "$location", "$http", "
         // TODO: notify if user passwords are not the same
         if (user.password == user.password2) {
             SignupFactory.post(user).$promise.then(function (data) {
-                console.log("Signup data ", data, data.auth);
                 $rootScope.currentUser = user;
             }, function (error) {
-                console.log("Signup error ", error);
             });
         }
     }
@@ -69,7 +60,6 @@ Harley.controller("SignUpCtrl", ["$scope", "$rootScope", "$location", "$http", "
 
 Harley.controller("LoginCtrl", ["$scope", "$rootScope", "$location", "$http", "LoginFactory", function($location, $scope, $http, $rootScope, LoginFactory) {
     $scope.login = function (user) {
-        console.log("LoginCtrl", user);
         LoginFactory.post(user);
     };
 }]);
